+++ comment_parser.py
